+++ fluent_contents/admin/genericextensions.py
@@ -55,11 +55,11 @@ class DynamicInlinesModelAdmin(ModelAdmin):
         return inline_instances
 
 
-    # Django 1.4:
-    # Inlines are created per request
+    # Django 1.4: Inlines are created per request
+    # Django 1.5: 'obj' parameter was added so it can be passed to 'has_change_permission' and friends.
 
-    def get_inline_instances(self, request):
-        inlines = super(DynamicInlinesModelAdmin, self).get_inline_instances(request)
+    def get_inline_instances(self, request, *args, **kwargs):
+        inlines = super(DynamicInlinesModelAdmin, self).get_inline_instances(request, *args, **kwargs)
         if self.extra_inlines_first:
             return self._get_extra_inline_instances() + inlines
         else: